+++ torch_points3d/metrics/colored_tqdm.py
@@ -1,5 +1,6 @@
 from tqdm.auto import tqdm
-from tqdm import std
+from collections import OrderedDict
+from numbers import Number
 import numpy as np
 
 from torch_points3d.utils.colors import COLORS
@@ -7,15 +8,15 @@ from torch_points3d.utils.colors import COLORS
 
 class Coloredtqdm(tqdm):
     def set_postfix(self, ordered_dict=None, refresh=True, color=None, round=4, **kwargs):
-        postfix = std._OrderedDict([] if ordered_dict is None else ordered_dict)
+        postfix = OrderedDict([] if ordered_dict is None else ordered_dict)
 
         for key in sorted(kwargs.keys()):
             postfix[key] = kwargs[key]
 
         for key in postfix.keys():
-            if isinstance(postfix[key], std.Number):
+            if isinstance(postfix[key], Number):
                 postfix[key] = self.format_num_to_k(np.round(postfix[key], round), k=round + 1)
-            if isinstance(postfix[key], std._basestring):
+            if isinstance(postfix[key], str):
                 postfix[key] = str(postfix[key])
             if len(postfix[key]) != round:
                 postfix[key] += (round - len(postfix[key])) * " "